requirements.txt
Django==4.0.3
djangorestframework==3.12.4
django-widget-tweaks==1.4.8
image==1.5.33
Pillow==9.1.0
pytest==7.1.1
pytz
asgiref==3.5.0
django-object-actions==4.0.0

stripe==5.0.0
django-environ==0.9.0
